Componentize ConsensusTable
