+++ components/Txns/ConsensusGroupV1.js
@@ -4,12 +4,16 @@ import Link from 'next/link'
 import animalHash from 'angry-purple-tiger'
 import TruncatedField from './TruncatedField'
 import { formatLocation } from '../Hotspots/utils'
+import ConsensusTable, {
+  makeArrayWorkWithAntTable,
+  generateColumns,
+} from '../ConsensusTable'
 
 import dynamic from 'next/dynamic'
 
 const ConsensusMapbox = dynamic(() => import('./ConsensusMapbox'), {
   ssr: false,
-  loading: () => <span style={{ height: '600px' }} />,
+  loading: () => <div style={{ minHeight: '600px' }} />,
 })
 
 const initialState = {
@@ -44,82 +48,20 @@ class ConsensusGroupV1 extends Component {
 
     return (
       <div>
-        <span>
-          {!loading ? (
-            <span style={{ height: '600px' }}>
+        <div>
+          {!loading && (
+            <div style={{ minHeight: '600px', backgroundColor: '#324b61' }}>
               <ConsensusMapbox members={consensusHotspots} />
-            </span>
-          ) : (
-            <div
-              style={{
-                height: '600px',
-                minHeight: '600px',
-                paddingLeft: '18px',
-                paddingTop: '18px',
-              }}
-            >
-              Loading...
             </div>
           )}
-        </span>
+        </div>
         <Descriptions bordered layout="vertical">
           <Descriptions.Item label="Consensus Group Members">
-            <ol
-              style={{ listStyle: 'none', paddingLeft: 0, paddingTop: '18px' }}
-            >
-              {consensusHotspots.map((m, index, { length }) => {
-                return (
-                  <li
-                    style={{
-                      paddingBottom: '18px',
-                    }}
-                  >
-                    <div
-                      style={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        justifyContent: 'flex-start',
-                      }}
-                    >
-                      <span
-                        style={{
-                          backgroundColor: '#ff6666',
-                          color: 'white',
-                          minHeight: '36px',
-                          height: '36px',
-                          minWidth: '36px',
-                          width: '36px',
-                          borderRadius: '36px',
-                          textAlign: 'center',
-                          boxShadow: '0 0 5px #cccccc',
-                          display: 'flex',
-                          alignItems: 'center',
-                          justifyContent: 'center',
-                          fontSize: '18px',
-                          marginBottom: '0px',
-                        }}
-                      >
-                        {index + 1}
-                      </span>
-                      <Link href={`/hotspots/${m.address}`}>
-                        <a style={{ paddingLeft: '10px', fontSize: '18px' }}>
-                          {animalHash(m.address)}
-                        </a>
-                      </Link>
-                    </div>
-                    <p
-                      style={{
-                        color: '#555',
-                        paddingLeft: 'calc(36px + 10px)',
-                      }}
-                    >
-                      {formatLocation(m.geocode)}
-                    </p>
-                  </li>
-                )
-              })}
-            </ol>
+            <ConsensusTable
+              loading={loading}
+              columns={generateColumns('current')}
+              dataSource={consensusHotspots}
+            />
           </Descriptions.Item>
         </Descriptions>
         <Descriptions bordered>
